Migracion a Short: por requisitos de MTU de BLE, el tamaño maximo de paquete(en principio) es de 20 bytes, por lo que modifico los dataclass de joystickControl y PidSettings a short para que entren, ya que en realidad ni siquiera estoy usando la precision adicional
Codigo funcionando con el ESP para transmitir paquetes

+++ app/src/main/java/com/example/hoverrobot/data/models/comms/AxisControl.kt
@@ -1,6 +1,6 @@
 package com.example.hoverrobot.data.models.comms
 
 data class AxisControl(
-    val axisX : Int,
-    val axisY : Int
+    val axisX: Short,
+    val axisY: Short
 )
